extract TTestTable into another file.
